Answer Arnaud's offline comment: move to gradient instead, except extract_dwi
