+++ scilpy/reconst/b_tensor_utils.py
@@ -6,8 +6,9 @@ from dipy.io.gradients import read_bvals_bvecs
 import nibabel as nib
 import numpy as np
 
-from scilpy.dwi.bvec_bval_tools import (normalize_bvecs, is_normalized_bvecs,
-                                        extract_dwi_shell)
+from scilpy.dwi.utils import extract_dwi_shell
+from scilpy.gradients.bvec_bval_tools import (normalize_bvecs,
+                                              is_normalized_bvecs)
 
 
 bshapes = {0: "STE", 1: "LTE", -0.5: "PTE", 0.5: "CTE"}
