+++ scripts/scil_compute_dti_metrics.py
@@ -44,8 +44,8 @@ from scilpy.io.utils import (add_overwrite_arg, assert_inputs_exist,
                              assert_outputs_exist, add_force_b0_arg)
 from scilpy.reconst.dti import convert_tensor_from_dipy_format, \
     supported_tensor_formats, tensor_format_description
-from scilpy.dwi.bvec_bval_tools import (normalize_bvecs, is_normalized_bvecs,
-                                        check_b0_threshold)
+from scilpy.gradients.bvec_bval_tools import (normalize_bvecs, is_normalized_bvecs,
+                                              check_b0_threshold)
 from scilpy.utils.filenames import add_filename_suffix, split_name_with_nii
 
 logger = logging.getLogger("Compute_DTI_Metrics")
