+++ scripts/scil_compute_memsmt_frf.py
@@ -40,6 +40,7 @@ import logging
 import nibabel as nib
 import numpy as np
 
+from scilpy.dwi.utils import extract_dwi_shell
 from scilpy.image.utils import extract_affine
 from scilpy.io.image import get_data_as_mask
 from scilpy.io.utils import (add_force_b0_arg,
@@ -47,7 +48,6 @@ from scilpy.io.utils import (add_force_b0_arg,
                              assert_inputs_exist, assert_outputs_exist,
                              assert_roi_radii_format)
 from scilpy.reconst.frf import compute_msmt_frf
-from scilpy.dwi.bvec_bval_tools import extract_dwi_shell
 from scilpy.reconst.b_tensor_utils import generate_btensor_input
