+++ scripts/scil_compute_msmt_frf.py
@@ -34,13 +34,13 @@ from dipy.io.gradients import read_bvals_bvecs
 import nibabel as nib
 import numpy as np
 
+from scilpy.dwi.utils import extract_dwi_shell
 from scilpy.io.image import get_data_as_mask
 from scilpy.io.utils import (add_force_b0_arg,
                              add_overwrite_arg, add_verbose_arg,
                              assert_inputs_exist, assert_outputs_exist,
                              assert_roi_radii_format)
 from scilpy.reconst.frf import compute_msmt_frf
-from scilpy.dwi.bvec_bval_tools import extract_dwi_shell
 
 
 def buildArgsParser():
