+++ scripts/scil_flip_gradients.py
@@ -9,8 +9,8 @@ import os
 
 from scilpy.io.utils import (add_overwrite_arg, assert_inputs_exist,
                              assert_outputs_exist)
-from scilpy.dwi.bvec_bval_tools import (flip_mrtrix_gradient_sampling,
-                                        flip_fsl_gradient_sampling)
+from scilpy.gradients.bvec_bval_tools import (flip_mrtrix_gradient_sampling,
+                                              flip_fsl_gradient_sampling)
 from scilpy.utils.util import str_to_index
