+++ scripts/scil_swap_gradient_axis.py
@@ -9,8 +9,8 @@ import os
 
 from scilpy.io.utils import (add_overwrite_arg, assert_inputs_exist,
                              assert_outputs_exist)
-from scilpy.dwi.bvec_bval_tools import (swap_fsl_gradient_axis,
-                                        swap_mrtrix_gradient_axis)
+from scilpy.gradients.bvec_bval_tools import (swap_fsl_gradient_axis,
+                                              swap_mrtrix_gradient_axis)
 from scilpy.utils.util import str_to_index
